# api/openapi.json
{
  "components": {
    "responses": {
...
      "NotFound": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`not_found`: no such resource in the caller's scope."
      },
...
      "Restricted": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`account_deletion_pending`: the account is scheduled for deletion and this session is export-only; `password_change_required`: the password was reset and must be changed with `PATCH /me` first."
      },
      "Unauthorized": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`unauthorized`: missing, invalid, expired or revoked access token (signed out, device removed, account disabled)."
      },
...
    },
    "schemas": {
...
      "CommitReverted": {
        "description": "Result of `POST /commits/{commit}/revert`.",
        "properties": {
          "commit": {
            "description": "The new commit (absent if nothing had to change).",
            "type": [
              "string",
              "null"
            ]
          },
          "paths": {
            "description": "Vault paths the revert changed, sorted.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "paths"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "Short-lived access token (PLAN §8).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Strata HTTP API. Every body is MessagePack (application/vnd.msgpack); errors are RFC 7807 problem details (application/problem+msgpack). See docs/WIRE_FORMAT.md.",
    "title": "Strata API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commits/{commit}/revert": {
      "post": {
        "operationId": "revert_commit",
        "parameters": [
          {
            "description": "Commit ID (40 hex characters).",
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/CommitReverted"
                }
              }
            },
            "description": "Reverted in one new commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`revert_conflict`: later changes conflict with the revert."
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Revert a whole commit (typically an `ai:` commit) on top of the current state.",
        "tags": [
          "notes"
        ]
      }
    },
...
    "/api/v1/notes": {
      "post": {
        "operationId": "create_note",
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            },
            "description": "Created; one `user: create <path>` commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`duplicate_candidates` (resend with `force`) or `path_taken`."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`invalid_name` (path) or `invalid_body`."
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Create a note (duplicate check unless `force`).",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/notes/by-path": {
      "get": {
        "operationId": "get_note_by_path",
        "parameters": [
          {
            "description": "Vault path, e.g. `notes/Pricing.md`.",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            },
            "description": "The note."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`not_found`."
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "A live note by vault path.",
        "tags": [
          "notes"
        ]
      }
    },
...
    "/api/v1/notes/{id}/backlinks": {
      "get": {
        "operationId": "get_backlinks",
        "parameters": [
          {
            "description": "Note ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Backlinks"
                }
              }
            },
            "description": "Backlinks."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Body links and relations pointing at a note, grouped by kind.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/notes/{id}/history": {
      "get": {
        "operationId": "get_note_history",
        "parameters": [
          {
            "description": "Note ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/History"
                }
              }
            },
            "description": "Commits that touched the note."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "The git history of a note, newest first.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/notes/{id}/history/{commit}": {
      "get": {
        "operationId": "get_note_revision",
        "parameters": [
          {
            "description": "Note ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          },
          {
            "description": "Commit ID (40 hex characters) from the note's history.",
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRevision"
                }
              }
            },
            "description": "The note in that commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "A note's content as of a commit in its history.",
        "tags": [
          "notes"
        ]
      }
    },
...
    "/api/v1/trash/{id}": {
      "delete": {
        "operationId": "purge_note",
        "parameters": [
          {
            "description": "Note ID (ULID) of a trashed note.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Purged; one `user: purge <path>` commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Permanently delete a trashed note and its sidecar (history keeps it).",
        "tags": [
          "notes"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}